moved pomodoro procedure to a new file. moved alerts to new file
